Refactor: 코드 정리
중복 코드 삭제

+++ src/main/java/com/project/simpleblog/service/CategoryServiceImpl.java
@@ -1,5 +1,6 @@
 package com.project.simpleblog.service;
 
+import com.project.simpleblog.domain.Board;
 import com.project.simpleblog.domain.Category;
 import com.project.simpleblog.domain.User;
 import com.project.simpleblog.dto.CategoryRequestDto;
@@ -7,6 +8,7 @@ import com.project.simpleblog.dto.CategoryResponseDto;
 import com.project.simpleblog.dto.StatusResponseDto;
 import com.project.simpleblog.exception.CategoryAlreadyExistsException;
 import com.project.simpleblog.exception.UnauthorizedBehaviorException;
+import com.project.simpleblog.repository.BoardRepository;
 import com.project.simpleblog.repository.CategoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -22,12 +24,14 @@ import java.util.stream.Collectors;
 public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final BoardRepository boardRepository;
 
     @Transactional
     @Override
     public CategoryResponseDto register(CategoryRequestDto categoryRequestDto, User user) {
         if (categoryRepository.findAllByUserIdOrderByCategory(user.getId()).stream().noneMatch(e -> e.getCategory().equals(categoryRequestDto.getCategory()))) {
-            return categoryRepository.save(new Category(categoryRequestDto, user.getId())).toResponseDto();
+            Category category = categoryRepository.save(new Category(categoryRequestDto, user.getId()));
+            return new CategoryResponseDto(category);
         }
         throw new CategoryAlreadyExistsException("이미 존재하는 카테고리입니다.");
     }
@@ -41,7 +45,7 @@ public class CategoryServiceImpl implements CategoryService {
     @Transactional(readOnly = true)
     @Override
     public List<CategoryResponseDto> getCategoriesByUser(User user) {
-        return categoryRepository.findAllByUserIdOrderByCategory(user.getId()).stream().map(Category::toResponseDto).collect(Collectors.toList());
+        return categoryRepository.findAllByUserIdOrderByCategory(user.getId()).stream().map(CategoryResponseDto::new).collect(Collectors.toList());
     }
 
     @Transactional
@@ -50,8 +54,10 @@ public class CategoryServiceImpl implements CategoryService {
         Category category = categoryRepository.findByUserIdAndCategory(user.getId(), categoryName).orElseThrow(() -> new NoSuchElementException("해당 카테고리는 존재하지 않습니다."));
 
         if (user.isAdmin() || user.isValidId(category.getUserId())) {
+            List<Board> boardList = boardRepository.findAllByUserIdAndCategoryOrderByCreatedAtDesc(user.getId(), categoryName);
             category.update(categoryRequestDto);
-            return category.toResponseDto();
+            boardList.forEach(board -> board.updateCategory(categoryRequestDto.getCategory()));
+            return new CategoryResponseDto(category);
         }
         throw new UnauthorizedBehaviorException("작성자만 수정할 수 있습니다.");
     }
@@ -60,10 +66,14 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public StatusResponseDto delete(String categoryName, User user) {
         Category category = categoryRepository.findByUserIdAndCategory(user.getId(), categoryName).orElseThrow(() -> new NoSuchElementException("해당 카테고리는 존재하지 않습니다."));
+        List<Board> boardList = boardRepository.findAllByUserIdAndCategoryOrderByCreatedAtDesc(user.getId(), categoryName);
 
         if (user.isAdmin() || user.isValidId(category.getUserId())) {
-            categoryRepository.delete(category);
-            return new StatusResponseDto("카테고리 삭제 성공", HttpStatus.OK.value());
+            if (boardList.size() == 0) {
+                categoryRepository.delete(category);
+                return new StatusResponseDto("카테고리 삭제 성공", HttpStatus.OK.value());
+            }
+            throw new IllegalArgumentException("카테고리를 삭제할 수 없습니다. 현재 게시글 수: " + boardList.size());
         }
         throw new UnauthorizedBehaviorException("작성자만 삭제할 수 있습니다.");
     }
